Sync latest: UI polish, match service hardening, utils module, backend seeds
Mobile:
- Add utils/image_utils.dart for shared image handling
- Profile screen enhancements, navigation scaffold updates
- Feed and challenge widgets refinements
- Match score screen cleanup
- Court service streamlining
- AppBar consistency across all screens

Backend:
- Match service spec + implementation hardening
- State discover script improvements
- Seed scripts for iconic/Las Vegas/South Bay runs
- Token helper page

+++ app/hooprank-starter-with-frontend/hooprank-starter/mobile/lib/models.dart
@@ -75,7 +75,7 @@ class User {
     final rawContestRate = json['contestRate'] ?? json['contest_rate'];
     final contestRate = rawContestRate != null
         ? _parseDouble(rawContestRate)
-        : (gamesPlayed > 0 ? gamesContested / gamesPlayed : 0.0);
+        : (gamesPlayed > 0 ? gamesContested.toDouble() / gamesPlayed : 0.0);
 
     // Handle both camelCase (app) and snake_case (production backend) field names
     final position = json['position']?.toString();
